leaderboard and subject ui added

+++ online_judge/leaderboard/views.py
@@ -21,12 +21,14 @@ from datetime import  datetime
 
 
 def connection():
-    return redis.StrictRedis(host='localhost', port=6379, db=0)
+    r = redis.StrictRedis(host=HOST, port=PORT, db=REDISDB)
+    return r
 
 
 @login_required()
 def set_leaderboard_subject(request):
-        r = redis.StrictRedis(host='localhost', port=6379, db=0)
+    try:
+        r = connection()
         subjectid = request.POST.get('subjectid')
         year = request.POST.get('year')
         #year = 2016
@@ -39,25 +41,29 @@ def set_leaderboard_subject(request):
         else:
             initialize_leaderboard_cache(request)
             return HttpResponseRedirect('/leaderboard/get_leaderboard')
-
+    except:
         return HttpResponseRedirect('/subject/')
 
-
-def flush_DB():
-    r = redis.StrictRedis(host='localhost', port=6379, db=0)
-    r.flushall()
-    return
+@login_required()
+def flush_RedisDB(request):
+    if request.user.is_superuser:
+        r = connection()
+        r.flushall()
+        return HttpResponseRedirect('/subject/')
+    else:
+        messages.add_message(request, messages.WARNING, 'You are not authorized!!')
+        return HttpResponseRedirect('/subject/all_subject')
 
 
 def initialize_leaderboard_cache(request):
-    r = redis.StrictRedis(host='localhost', port=6379, db=0)
+    r = connection()
     subjectid = request.session.get('leaderboard_subjectid')
     subject = Subject.objects.get(id=subjectid)
     year = request.session.get('leaderboard_year')
     leaderboard = Leaderboard.objects.filter(subject=subject, year=year).values('subject', 'week', 'student').annotate(Sum('maxscore'))
 
     student_list = set()
-    week_list = Week.objects.filter(subject=subject)
+    week_list = Week.objects.filter(subject=subject, isdeleted=False)
     for l in leaderboard:
         student_list.add(l['student'])
     subject_rank_name = "rank:"+str(year)+':'+str(subjectid)
@@ -74,8 +80,8 @@ def initialize_leaderboard_cache(request):
     r.hset("ranklist", subject_rank_name, subject_rank_name)
     r.hset("leaderboard", str(year)+':'+str(subjectid), str(year)+':'+str(subjectid))
 
-    r.expire(subject_rank_name, 400000)
-    r.expire(str(year)+':'+str(subjectid), 400000)
+    r.expire(subject_rank_name, 100000)
+    r.expire(str(year)+':'+str(subjectid), 100000)
 
     for l in leaderboard:
         name = str(year)+':'+str(l['subject'])+':'+str(l['student'])
@@ -96,106 +102,104 @@ def initialize_leaderboard_cache(request):
 
 @login_required()
 def get_leaderboard(request):
-    subjectid = request.session.get('leaderboard_subjectid')
-    subject = Subject.objects.get(id=subjectid)
-    year = request.session.get('leaderboard_year')
-    weeks = Week.objects.filter(subject=subject)
-    week = {}
-    for w in weeks:
-        week[w.id] = w.name
+    try:
+        r = connection()
+        subjectid = request.session.get('leaderboard_subjectid')
+        subject = Subject.objects.get(id=subjectid)
+        year = request.session.get('leaderboard_year')
+        weeks = Week.objects.filter(subject=subject, isdeleted=False)
+        week = {}
+        for w in weeks:
+            week[w.id] = w.name
 
-    students =Student.objects.filter(year=year)
-    student = {}
-    for s in students:
-        student[str(s.id)] = s.user.username
+        students =Student.objects.filter(year=year)
+        student = {}
+        for s in students:
+            student[str(s.id)] = s.user.username
 
-    r = redis.StrictRedis(host='localhost', port=6379, db=0)
-    data = {}
-    ranklist_data = {}
 
-    #d = datetime.now()
-    #print(d.microsecond)
+        data = {}
+        ranklist_data = {}
 
-    ranklist = r.zrevrange("rank:"+str(year)+':'+str(subjectid), 0, -1, withscores=True)
-    for ra in ranklist:
-        ra = list(ra)
-        key = str(ra[0])
-        value = ra[1]
-        key = list(key.split(':'))[-1]
-        key = str(key)
-        key = key[:len(key)-1]
-        ranklist_data[key] = int(value)
+        #d = datetime.now()
+        #print(d.microsecond)
 
-    subject_key = str(year)+':'+str(subjectid)
-    for r_student in r.hgetall(subject_key):
-
-        temp = r.hgetall(r_student)
-        t = {}
-        outer_key = list(str(r_student).split(':'))[-1]
-        outer_key = outer_key[:len(outer_key)-1]
-        for key in temp:
-            value = temp[key]
-            key = str(key)
+        ranklist = r.zrevrange("rank:"+str(year)+':'+str(subjectid), 0, -1, withscores=True)
+        for ra in ranklist:
+            ra = list(ra)
+            key = str(ra[0])
+            value = ra[1]
             key = list(key.split(':'))[-1]
+            key = str(key)
             key = key[:len(key)-1]
+            ranklist_data[key] = int(value)
 
-            value = int(value)
-            t[key] = value
-            #print(value)
-        data[outer_key] = t
+        subject_key = str(year)+':'+str(subjectid)
+        for r_student in r.hgetall(subject_key):
 
-    data_leaderboard = {}
-    count =1
-    for key in ranklist_data:
-        temp_dic={}
-        temp_dic['student_name'] = student[key]
-        temp_dic['totalscore'] = ranklist_data[key]
+            temp = r.hgetall(r_student)
+            t = {}
+            outer_key = list(str(r_student).split(':'))[-1]
+            outer_key = outer_key[:len(outer_key)-1]
+            for key in temp:
+                value = temp[key]
+                key = str(key)
+                key = list(key.split(':'))[-1]
+                key = key[:len(key)-1]
 
-        student_dic = data[key]
-        for student_d in student_dic:
-            temp_dic[student_d] = student_dic[student_d]
-        data_leaderboard[count] = temp_dic
-        count += 1
-        #print(temp_dic)
+                value = int(value)
+                t[key] = value
+            data[outer_key] = t
 
-    #d = datetime.now()
-    #print(d.microsecond)
-    c = {}
-    c['subject'] = subject
-    c['data'] = data_leaderboard
-    c['weeks'] = week
-    c['students'] = student
-    c['ranklist'] = ranklist_data
+        data_leaderboard = {}
+        count =1
+        for key in ranklist_data:
+            temp_dic={}
+            temp_dic['student_name'] = student[key]
+            temp_dic['totalscore'] = ranklist_data[key]
 
-    return render(request,'leaderboard/leaderboard.html', c)
+            student_dic = data[key]
+            for student_d in student_dic:
+                temp_dic[student_d] = student_dic[student_d]
+            data_leaderboard[count] = temp_dic
+            count += 1
 
+        #d = datetime.now()
+        #print(d.microsecond)
+        c = {}
+        c['subject'] = subject
+        c['data'] = data_leaderboard
+        c['weeks'] = week
+        c['students'] = student
+        c['ranklist'] = ranklist_data
+
+        return render(request,'leaderboard/leaderboard.html', c)
+    except:
+        messages.add_message(request, messages.WARNING, 'something wrong!!')
+        return HttpResponseRedirect('/subject/all_subject')
 
 def update_cache(leaderboard):
-    r = redis.StrictRedis(host='localhost', port=6379, db=0)
+    r = connection()
     student_id = leaderboard.student.id
     subject_id = leaderboard.subject.id
     week_id = leaderboard.week.id
     year = leaderboard.year
     key = str(year)+':'+str(subject_id)+":"+str(student_id)+":"+str(week_id)
-
     name = str(year)+':'+str(subject_id)+":"+str(student_id)
     oldval = int(r.hget(name, key))
     incre_val = leaderboard.maxscore - oldval
     r.hincrby(name, key, incre_val)
-
     name_for_z = "rank:"+str(year)+':' + str(subject_id)
     key_for_z = "rank:" + str(year)+':' + str(subject_id) + ":" + str(student_id)
-
     r.zincrby(name_for_z, incre_val, key_for_z)
 
     return
 
 
 def update_cache_week(subjectid):
-    r = redis.StrictRedis(host='localhost', port=6379, db=0)
+    r = connection()
     ymax = Student.objects.all().aggregate(Max('year'))['year__max']
     ymin = Student.objects.all().aggregate(Min('year'))['year__min']
-
     for year in range(ymin,ymax+1):
         hash_key = str(year)+':'+str(subjectid)
         set_key = "rank:"+str(year)+':'+str(subjectid)
@@ -218,9 +222,7 @@ def show_leaderboard(request):
     for week in weeks:
         week_dic[week.id] = week.name
     c = {}
-
     leaderboard = Leaderboard.objects.filter(subject=subject,year=year).values('subject', 'student','week').annotate(Sum('maxscore'))
-
     students = Leaderboard.objects.filter(subject=subject, year=year)
     # print(leaderboard)
     leaderboard_data = {}
@@ -233,20 +235,15 @@ def show_leaderboard(request):
         for week in weeks:
             t[week.id] = 0
         leaderboard_data[student1.student.id] = t
-    #print(leaderboard)
     for l in leaderboard:
         if int(l['subject']) == int(subjectid):
             leaderboard_data[l['student']][l['week']] = l['maxscore__sum']
             leaderboard_data[l['student']]['total'] += l['maxscore__sum']
-    #print(leaderboard_data)
     data = dict(sorted(leaderboard_data.items(), key=lambda x: x[1]['total'], reverse=True))
-    #print(data)
     c={}
     c['leaderboard_data'] = data
     c['subject'] = subject
     c['weeks'] = week_dic
-    #initialize_leaderboard_cache(request)
-    #get_leaderboard(request)
     #d = datetime.now()
     #print(d.microsecond)
